# docs/40-api/openapi.yaml
openapi: 3.1.0
info:
  title: AI PPT Builder API
  version: 0.1.0
  description: |
    专注 frontend 的 PPT Coding Agent 的后端契约。
    REST 用于控制/查询，SSE 用于流式输出（见 /runs/{id}/events）。
    无鉴权（单用户单机，默认监听回环）。
servers:
  - url: http://127.0.0.1:8787/api/v1
    description: 本地开发
...
paths:
...
  /assets:
    get:
      tags: [assets]
      summary: 列出个人仓库资产（含预置与用户新增）
      parameters:
        - { name: kind, in: query, required: false, schema: { type: string, enum: [layout, component, theme, fx] } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Asset' }
        '400': { $ref: '#/components/responses/Error' }
    post:
      tags: [assets]
      summary: 新增资产（提交 manifest + 载荷，服务端校验资产协议 schema）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAssetRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Asset' }
        '409': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
  /assets/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      tags: [assets]
      summary: 获取资产
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Asset' }
        '404': { $ref: '#/components/responses/Error' }
    patch:
      tags: [assets]
      summary: 修改资产（载荷/元数据；产生资产版本）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PatchAssetRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Asset' }
        '404': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
    delete:
      tags: [assets]
      summary: 删除资产
      responses:
        '204': { description: No Content }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
...
components:
  responses:
    Error:
      description: 统一错误响应
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object }
...
    Asset:
      type: object
      required: [id, name, kind, version, source, manifest_path, dir, created_at]
      properties:
        id: { type: string }
        name: { type: string }
        kind: { type: string, enum: [layout, component, theme, fx] }
        version: { type: string }
        source: { type: string, enum: [preset, user] }
        description: { type: string }
        tags: { type: array, items: { type: string } }
        manifest_path: { type: string }
        dir: { type: string }
        created_at: { type: integer }
    CreateAssetRequest:
      type: object
      required: [manifest, payload]
      properties:
        manifest:
          type: object
          description: '统一信封（含 name/kind/version/description 等），须符合 asset-manifest schema'
        payload:
          type: object
          description: '按 kind 对应载荷：html/css（layout/component）、tokens（theme）、js（fx）'
          additionalProperties: { type: string }
    PatchAssetRequest:
      type: object
      description: '修改资产载荷或元数据；采用锚定文本替换或整体替换载荷'
      properties:
        manifest: { type: object }
        payload:
          type: object
          additionalProperties: { type: string }
        edits:
          type: array
          description: '锚定文本替换（与 patch_slide 同构）'
          items:
            type: object
            required: [file, old_text, new_text]
            properties:
              file: { type: string }
              old_text: { type: string }
              new_text: { type: string }
...